fix UI circular dependencies
