--- src/UI/IconState/types.ts
@@ -1,6 +1,0 @@
-export interface Props {
-  icon: any;
-  className?: string;
-  iconSize?: '2x' | '3x' | '5x';
-  shouldRenderDefaultCss?: boolean;
-}
