--- src/UI/Progress/type.ts
@@ -1,10 +1,0 @@
-export interface Props {
-  id: string;
-  done: boolean;
-  children: React.ReactNode;
-  expiresIn?: number;
-  progress: {
-    startTime: number;
-    endTime: number;
-  };
-}
