--- src/UI/TransactionsToastList/types.ts
@@ -1,11 +1,0 @@
-import { SignedTransactionsType } from 'types';
-
-export interface TransactionsToastListPropsType {
-  toastProps?: any;
-  className?: string;
-  withTxNonce?: boolean;
-  shouldRenderDefaultCss?: boolean;
-  pendingTransactions?: SignedTransactionsType;
-  signedTransactions?: SignedTransactionsType;
-  successfulToastLifetime?: number;
-}
