--- src/UI/extension/LoginButton/types.ts
@@ -1,12 +1,0 @@
-import { ReactNode } from 'react';
-
-export interface ExtensionLoginButtonPropsType {
-  token?: string;
-  children?: ReactNode;
-  className?: string;
-  buttonClassName?: string;
-  callbackRoute: string;
-  loginButtonText?: string;
-  redirectAfterLogin?: boolean;
-  shouldRenderDefaultCss?: boolean;
-}
