+++ src/UI/index.tsx
@@ -5,7 +5,6 @@ import LedgerLoginButton from './ledger/LoginButton';
 import LedgerLoginContainer from './ledger/LoginModal';
 
 import NotificationModal from './NotificationModal';
-import * as DappCorePages from './pages';
 import PageState from './PageState';
 import ProgressSteps from './ProgressSteps';
 import SignTransactionsModals from './SignTransactionsModals';
@@ -39,7 +38,5 @@ export default {
   WebWalletLoginButton,
   Trim,
   UsdValue,
-  ProgressSteps,
-  DappCorePages,
-
+  ProgressSteps
 };
