--- src/UI/walletConnect/WalletConnectLoginButton/types.ts
@@ -1,19 +1,0 @@
-import { ReactNode } from 'react';
-
-export interface WalletConnectLoginButtonPropsType {
-  onModalOpens?: (props?: any) => void;
-  onModalCloses?: (props?: any) => void;
-  children?: ReactNode;
-  lead?: string;
-  title?: string;
-  className?: string;
-  logoutRoute?: string;
-  callbackRoute: string;
-  loginButtonText?: string;
-  buttonClassName?: string;
-  shouldRenderDefaultCss?: boolean;
-  wrapContentInsideModal?: boolean;
-  redirectAfterLogin?: boolean;
-  hideButtonWhenModalOpens?: boolean;
-  token?: string;
-}
