--- src/UI/walletConnect/WalletConnectLoginContainer/types.ts
@@ -1,13 +1,0 @@
-export interface LoginModalPropsType {
-  lead?: string;
-  title?: string;
-  className?: string;
-  logoutRoute?: string;
-  callbackRoute: string;
-  loginButtonText: string;
-  wrapContentInsideModal?: boolean;
-  shouldRenderDefaultCss?: boolean;
-  redirectAfterLogin?: boolean;
-  token?: string;
-  onClose?: () => void;
-}
