--- src/UI/webWallet/LoginButton/types.ts
@@ -1,11 +1,0 @@
-import { ReactNode } from 'react';
-
-export interface WebWalletLoginButtonPropsType {
-  token?: string;
-  className?: string;
-  callbackRoute: string;
-  buttonClassName?: string;
-  children?: ReactNode;
-  loginButtonText?: string;
-  shouldRenderDefaultCss?: boolean;
-}
