+++ src/types/UI.ts
@@ -1,3 +1,5 @@
+import { ReactNode } from 'react';
+
 export interface DenominateType {
   value: string;
   showLastNonZeroDecimal?: boolean;
@@ -7,4 +9,15 @@ export interface DenominateType {
   denomination?: number;
   egldLabel?: string;
   'data-testid'?: string;
+}
+
+export interface ExtensionLoginButtonPropsType {
+  token?: string;
+  children?: ReactNode;
+  className?: string;
+  buttonClassName?: string;
+  callbackRoute: string;
+  loginButtonText?: string;
+  redirectAfterLogin?: boolean;
+  shouldRenderDefaultCss?: boolean;
 }
